\name{rJava-package}
\alias{rJava-package}
\alias{rJava}
\docType{package}
\title{
\packageTitle{rJava}
}
\description{
\packageDescription{rJava}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rJava}
\packageIndices{rJava}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rJava}

Maintainer: \packageMaintainer{rJava}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
